Create file structure for restaurant feature. Extract ui to RestaurantScreen.
