+++ src/features/restaurants/components/RestaurantInfoComponent.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function RestaurantInfo() {
+  return <div>RestaurantInfo</div>;
+}
