+++ src/features/restaurants/screens/RestaurantsScreen.tsx
@@ -1,0 +1,41 @@
+import React, { useState } from "react";
+import { StatusBar } from "react-native";
+import { Searchbar } from "react-native-paper";
+import { StyleSheet, Text, View, SafeAreaView } from "react-native";
+
+export default function RestaurantsScreen() {
+  const [searchQuery, setSearchQuery] = useState("");
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <View style={styles.searchContainer}>
+        <Searchbar
+          placeholder="Search"
+          onChangeText={setSearchQuery}
+          value={searchQuery}
+          elevation={2}
+        />
+      </View>
+      <View style={styles.listContainer}>
+        <Text>List area</Text>
+      </View>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    marginTop: StatusBar.currentHeight,
+  },
+  searchContainer: {
+    backgroundColor: "green",
+    padding: 20,
+  },
+  listContainer: {
+    flex: 1,
+    backgroundColor: "blue",
+    padding: 20,
+  },
+});
